+++ CHANGELOG.md
@@ -10,6 +10,15 @@ Boca adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 
+### Added
+
+- `version` command, also accessible using `-v`, `-V` and `--version` flags.
+- `init:custom` command.
+- Custom commands:
+    - Loaded from the `BOCA_CUSTOM_COMMANDS` file, defaulting to `boca.py` in the current directory.
+    - All `@click.command()` or `@click.group()` items loaded in the file are
+    picked up and registered on the CLI.
+
 ## v0.0.1 - 2018-01-20
 
 Initial release to PyPI.